requirements.txt
fastapi
uvicorn
openai
python-multipart
python-dotenv
requests
pillow
pillow-heif